feat(networkd): Adding support for reloading the networkd config
Using the network-convert and config-drive options instead of the file option.
This will parse user_data as before in addition to parsing the network configuration
and dumping the result into systemd's runtime network directory. After generating
the configs, the network interfaces are downed, the 8012q LKM is loaded, and
networkd is restarted, loading the new network configuration.

Ideally, the network config will be generated before networkd is started,
eliminating the need for this restart hack.